requirements.txt
streamlit
google-generativeai

 

